redesign router macro
