+++ packages/storey-macros/src/lib.rs
@@ -1,20 +1,22 @@
 mod key_derive;
-mod storable;
+mod router;
 
-#[proc_macro_derive(Storable, attributes(key))]
-pub fn storable_derive(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+use proc_macro::TokenStream;
+
+#[proc_macro_attribute]
+pub fn router(_attrs: TokenStream, input: TokenStream) -> TokenStream {
     let input = syn::parse_macro_input!(input as syn::ItemStruct);
 
-    let expanded = match storable::derive(input) {
+    let expanded = match router::entry(input) {
         Ok(res) => res,
         Err(e) => e.into_compile_error(),
     };
 
-    proc_macro::TokenStream::from(expanded)
+    TokenStream::from(expanded)
 }
 
 #[proc_macro_derive(Key)]
-pub fn key_derive(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+pub fn key_derive(input: TokenStream) -> TokenStream {
     let input = syn::parse_macro_input!(input as syn::ItemStruct);
 
     let expanded = match key_derive::key_derive(input) {
@@ -22,11 +24,11 @@ pub fn key_derive(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
         Err(e) => e.into_compile_error(),
     };
 
-    proc_macro::TokenStream::from(expanded)
+    TokenStream::from(expanded)
 }
 
 #[proc_macro_derive(OwnedKey)]
-pub fn owned_key_derive(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+pub fn owned_key_derive(input: TokenStream) -> TokenStream {
     let input = syn::parse_macro_input!(input as syn::ItemStruct);
 
     let expanded = match key_derive::owned_key_derive(input) {
@@ -34,5 +36,5 @@ pub fn owned_key_derive(input: proc_macro::TokenStream) -> proc_macro::TokenStre
         Err(e) => e.into_compile_error(),
     };
 
-    proc_macro::TokenStream::from(expanded)
+    TokenStream::from(expanded)
 }
