+++ packages/storey-macros/src/router/def.rs
@@ -1,0 +1,10 @@
+pub struct RouterDef {
+    pub name: syn::Ident,
+    pub fields: Vec<Field>,
+}
+
+pub struct Field {
+    pub ty: syn::Type,
+    pub name: syn::Ident,
+    pub key: u8,
+}
