+++ packages/storey-macros/src/router/parse/mod.rs
@@ -1,0 +1,7 @@
+use syn::ItemStruct;
+
+use super::def::RouterDef;
+
+pub fn parse(_input: ItemStruct) -> Result<RouterDef, syn::Error> {
+    todo!()
+}
